moved eventDataFactory init into params out of constructor for better testability
