+++ collector-bitmovin-player/src/main/java/com/bitmovin/analytics/bitmovin/player/BitmovinPlayerCollector.java
@@ -1,9 +1,16 @@
 package com.bitmovin.analytics.bitmovin.player;
 
 import android.content.Context;
+import android.content.pm.ApplicationInfo;
+import android.content.pm.PackageInfo;
+import android.content.pm.PackageManager;
+import android.os.Build;
 
 import com.bitmovin.analytics.BitmovinAnalytics;
 import com.bitmovin.analytics.BitmovinAnalyticsConfig;
+import com.bitmovin.analytics.data.DeviceInformationProvider;
+import com.bitmovin.analytics.data.EventDataFactory;
+import com.bitmovin.analytics.data.UserIdProvider;
 import com.bitmovin.player.BitmovinPlayer;
 
 public class BitmovinPlayerCollector extends BitmovinAnalytics {
@@ -25,7 +32,8 @@ public class BitmovinPlayerCollector extends BitmovinAnalytics {
     }
 
     public void attachPlayer(BitmovinPlayer player) {
-        BitmovinSdkAdapter adapter = new BitmovinSdkAdapter(player, this.bitmovinAnalyticsConfig, this.context,
+        EventDataFactory factory = new EventDataFactory(this.bitmovinAnalyticsConfig, context, new DeviceInformationProvider(context, getUserAgent(context)), new UserIdProvider(context));
+        BitmovinSdkAdapter adapter = new BitmovinSdkAdapter(player, this.bitmovinAnalyticsConfig, factory,
                 this.playerStateMachine);
 
         this.attach(adapter);
@@ -34,5 +42,24 @@ public class BitmovinPlayerCollector extends BitmovinAnalytics {
             BitmovinSdkAdAdapter adAdapter = new BitmovinSdkAdAdapter(player, this.adAnalytics);
             this.attachAd(adAdapter);
         }
+    }
+
+    private String getUserAgent(Context context) {
+        ApplicationInfo applicationInfo = context.getApplicationInfo();
+        int stringId = applicationInfo.labelRes;
+        String applicationName = "Unknown";
+        if (stringId == 0 && applicationInfo.nonLocalizedLabel != null) {
+            applicationInfo.nonLocalizedLabel.toString();
+        }
+        String versionName;
+        try {
+            String packageName = context.getPackageName();
+            PackageInfo info = context.getPackageManager().getPackageInfo(packageName, 0);
+            versionName = info.versionName;
+        } catch (PackageManager.NameNotFoundException var5) {
+            versionName = "?";
+        }
+
+        return applicationName + "/" + versionName + " (Linux;Android " + Build.VERSION.RELEASE + ") " + "BitmovinPlayer/" + BitmovinUtil.getPlayerVersion();
     }
 }
